requirements.txt
alibabacloud-credentials==0.1.4
alibabacloud-dysmsapi20170525==2.0.2
alibabacloud-endpoint-util==0.0.3
alibabacloud-openapi-util==0.1.4
alibabacloud-tea==0.2.6
alibabacloud-tea-console==0.0.1
alibabacloud-tea-openapi==0.2.5
alibabacloud-tea-util==0.3.4
aliyun-python-sdk-core-v3==2.13.32
amqp==5.0.6
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.5
django-celery-beat==2.2.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-formtools==2.3
django-haystack==3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-redis==4.12.1
django-reversion==3.0.9
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dnspython==1.16.0
elasticsearch==2.4.1
et-xmlfile==1.1.0
eventlet==0.30.2
# fdfs-client-py @ file:///D:/Download/Django%E9%98%B6%E6%AE%B5%E8%B5%84%E6%96%99%E5%88%86%E7%A6%BB%E7%89%88/Django%E9%98%B6%E6%AE%B5%E8%B5%84%E6%96%99/%E7%BE%8E%E5%A4%9A%E5%95%86%E5%9F%8E%E9%A1%B9%E7%9B%AE-%E7%AC%AC06%E5%A4%A9/02-%E5%85%B6%E4%BB%96%E8%B5%84%E6%96%99/fdfs_client-py-master.zip
future==0.18.2
gevent==21.1.2
greenlet==1.0.0
httplib2==0.9.2
idna==2.10
importlib-metadata==3.10.1
itsdangerous==1.1.0
jmespath==0.10.0
kombu==5.0.2
MarkupPy==1.14
mutagen==1.40.0
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
prompt-toolkit==3.0.18
pycparser==2.20
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==3.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
QQLoginTool==0.3.0
redis==3.5.3
requests==2.25.1
ronglian-sms-sdk==1.0.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
# xadmin @ file:///D:/Download/sshwsfc-xadmin-0.4.0-378-g291b63b/sshwsfc-xadmin-291b63b.zip
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
